fix: address all CodeRabbit review items on PR #58
Major fixes:
- Sentinel ErrConfigNotFound + context.WithTimeout on FetchConfig (M1)
- Deterministic service ordering via sorted map keys (M2)
- Sanitize .URL/.BaseURL/.HeaderValue in Markdown templates (M3)
- Handle 404 on PUT with retry-as-POST for deleted comments (M4)
- Bound io.ReadAll with 1MB LimitReader in config fetchers (M5)
- Strict port parsing with strconv.ParseInt, fix registry:port (M6)
- runPreviewStatus accepts io.Writer, use cmd.OutOrStdout() (M7)
- CI: require KUBECONFIG, pin kubectl version (M8)
- CI: fix apiVersion, commitSHA typo, quote TTL, validate services (M9)
- CI: fix component path reference in README (M10)
- Fix namespace auto-discovery doc claim (M11)
- Add CommentID to proto PreviewGroupStatus (M12)
- Register PreviewGroup validating webhook (M13)
- Consolidate reconcileHTTPRoute/reconcileGRPCRoute into reconcileRoute (D1)
- Wire PreviewGroupNotifier lifecycle calls in reconciler (D2)

Minor fixes:
- Default image uses registry.gitlab.com prefix (N2)
- Remove hard-coded preview.example.com placeholder (N3)
- Sentinel errors ErrNotGitLabSource/ErrMissingMergeRequest (N4)
- Assert parsed fields in CLI tests, not just count (N5)
- Fix --services → --service in help text (N6)
- Wire signal.NotifyContext for SIGINT/SIGTERM (N7)
- AlreadyExists continues loop instead of early return (N8)
- Validate label value + tighten child selection with managed-by (N9)

+++ api/v1alpha1/previewgroup_types.go
@@ -43,7 +43,7 @@ const (
 )
 
 // PreviewGroupServiceSpec defines a single service in a preview group.
-// Only Name is required; Namespace, Port, ParentRef, and PathPrefix
+// Only Name is required; Port, ParentRef, and PathPrefix
 // are auto-discovered from existing Kubernetes Services if omitted.
 type PreviewGroupServiceSpec struct {
 	// Name of the Kubernetes Service to preview. Must match an existing Service.
@@ -68,7 +68,7 @@ type PreviewGroupServiceSpec struct {
 	// +optional
 	Endpoint string `json:"endpoint,omitempty"`
 
-	// Namespace is the target namespace. Auto-discovered from existing Service if empty.
+	// Namespace is the Kubernetes namespace where the service will be deployed. Required. If omitted, falls back to the controller's default namespace.
 	// +optional
 	Namespace string `json:"namespace,omitempty"`
 
@@ -250,9 +250,9 @@ type PreviewGroupStatus struct {
 	// +optional
 	Conditions []metav1.Condition `json:"conditions,omitempty"`
 
-	// CommentID stores the ID of the MR/PR comment.
+	// CommentID is the ID of the comment tracking the preview group status.
 	// +optional
-	CommentID int `json:"commentID,omitempty"`
+	CommentID int64 `json:"commentID,omitempty"`
 }
 
 // +kubebuilder:object:root=true
